libconsole: Really allow to use /dev/console as a fallback in showconsole
consalloc() always returns an error when it is called from
showconsole (with io = 0). As a result, /dev/console is ignored
because it is not assigned to the global "cons" pointer
in getconsoles().

consalloc() should return success (1) when the console is added
to the list and the io-related operations are skipped.

Signed-off-by: Petr Mladek <pmladek@suse.com>