libconsole: Add console into the list only when successfully allocated
consalloc() initializes the console when it is used from blogd (io = 1).
This operation might fail when the related tty device cannot be opened.
The structure is freed in this case but it was already added into
the list before. This might cause access to a freed memory.

This patch moves the initialization code into a separate function.
The structure is added to the list only when the initialization
succeeds. It is freed otherwise.

Signed-off-by: Petr Mladek <pmladek@suse.com>